Homomorphic encryption code checkin
HECrypto1 is trying to be the origional work by Craig Gentry, but
instead of encoding things in floats, which are bools with error, it's
encoding as bits without error.  A simplification for educational and
investigatory purposes.
HECrypto2 is an attempt at Craig Gentry's first work w/o bootstrapping,
to make a leveled he.